pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dropbox-link-generate"
version = "0.1.0"
description = "A command-line tool to generate Dropbox sharing links"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Dropbox Link Generate", email = "niceday@example.com"},
]
keywords = ["dropbox", "cli", "sharing", "links"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
    "Topic :: Internet :: File Transfer Protocol (FTP)",
]

dependencies = [
    "dropbox>=11.0.0",
    "requests>=2.28.0",
    "click>=8.0.0",
    "pyperclip>=1.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
dplk = "dropbox_link_generate.cli:main"

[project.urls]
Homepage = "https://github.com/niceday/dropbox-link-generate"
Repository = "https://github.com/niceday/dropbox-link-generate.git"
Issues = "https://github.com/niceday/dropbox-link-generate/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=dropbox_link_generate --cov-report=term-missing"